+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - 新命令 -bpm
 - 修改了 stat 绘制的样式并增加了 tth 信息
 - 改善了 helper 的绘制
+- 增加了 ContentMessage 的 content 属性，在读取和编辑信息时推荐使用此属性
 
 ## Bugs
 - 修复了其他三模式 accuracy 计算错误的问题